@@ -9,14 +9,18 @@
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
+  "engines": {
+    "node": "10"
+  },
   "main": "lib/index.js",
   "dependencies": {
-    "firebase-admin": "~7.0.0",
-    "firebase-functions": "^2.2.0"
+    "firebase-admin": "8.3.0",
+    "firebase-functions": "^3.2.0"
   },
   "devDependencies": {
+    "@types/firebase": "^3.2.1",
     "tslint": "^5.12.0",
-    "typescript": "^3.2.2"
+    "typescript": "^3.5.3"
   },
   "private": true
 }
