@@ -9,9 +9,11 @@ import { Observable, of, fromEvent } from "rxjs";
 import { Extended, User, StoreInfo } from "./interfaces/data-models";
 import { UsersService } from "./providers/AppData/users.service";
 import { map } from "rxjs/internal/operators/map";
-import { switchMap, combineLatest } from "rxjs/operators";
+import { switchMap, combineLatest, tap } from "rxjs/operators";
 import { StoreInfoService } from "./providers/AppData/store-info.service";
 import { Router } from "@angular/router";
+import { Location } from "@angular/common";
+// import firebase from "firebase";
 
 @Component({
   selector: "app-root",
@@ -32,11 +34,14 @@ export class AppComponent implements OnInit, OnDestroy {
     private storesService: StoreInfoService,
     private activeStoreService: ActiveStoreService,
     private router: Router,
+    private location: Location,
     private modalCtrl: ModalController
   ) {
     this.initializeApp();
 
-    this.user$ = this.usersService.currentUser;
+    this.user$ = this.usersService.currentUser.pipe(tap(u => {
+      console.log(u);
+    }));
     //  this.authService.user.pipe(
     //   switchMap(u => {
     //     return u ? this.usersService.get(u.uid) : of(null);
@@ -54,31 +59,33 @@ export class AppComponent implements OnInit, OnDestroy {
 
     this.authService.user.pipe(combineLatest(this.activeStore$)).subscribe(([user, store]) => {
       if (!user) {
-        this.router.navigateByUrl("/login");
-      } else if ( store ) {
+        // this.location.
+        this.router.navigateByUrl("/login", { replaceUrl: true });
+      } else if (store) {
         if (!this.initialAppLoad) {
-          this.router.navigateByUrl("/");
+          this.router.navigateByUrl("StoreBase/ProductsList", { replaceUrl: true });
           this.initialAppLoad = false;
         }
-       } else {
+      } else {
         this.router.navigateByUrl("/UserStores");
         this.initialAppLoad = false;
-     }
+      }
     });
   }
   ngOnInit() {
-    const event = fromEvent(document, 'backbutton');
-    this.backbuttonSubscription = event.subscribe(async () => {
-        const modal = await this.modalCtrl.getTop();
-        if (modal) {
-            modal.dismiss();
-        }
-    });
-}
+    // const event = fromEvent(document, 'backbutton');
+    // this.backbuttonSubscription = event.subscribe(async () => {
+    //   const modal = await this.modalCtrl.getTop();
+    //   if (modal) {
+    //     modal.dismiss();
+    //   }
+    // });
+    // firebase.firestore.setLogLevel("debug");
+  }
 
-ngOnDestroy() {
+  ngOnDestroy() {
     this.backbuttonSubscription.unsubscribe();
-}
+  }
   initializeApp() {
     this.platform.ready().then(() => {
       this.statusBar.styleDefault();
