@@ -6,7 +6,7 @@ import { ActiveStoreService } from "../AppData/active-store.service";
 import { StorePathConfig } from "../../interfaces/StorePathConfig";
 import { ImageService } from "../Image/image.service";
 import { combineLatest, forkJoin, Observable } from "rxjs";
-import { finalize, map, last, switchMap, catchError } from "rxjs/operators";
+import { finalize, map, last, switchMap, catchError, tap } from "rxjs/operators";
 
 @Injectable({
   providedIn: "root"
@@ -22,10 +22,10 @@ export class ImagesDataService extends StoreDataService<ImageFile> {
   ) {
     super(afs, activeStoreService, StorePathConfig.Images);
     console.log("Hello ImagesFsRepository Provider");
-    this.extendedList = this.extendList(this.List());
+    this.extendedList = this.extendList(this.list);
   }
   private extendList(list: Observable<Extended<ImageFile>[]>) {
-    return list.pipe(
+    return this.list.pipe(
       map(extImages => {
         return extImages.map(extImage => {
           if (!extImage.data.isUploaded) {
@@ -42,7 +42,7 @@ export class ImagesDataService extends StoreDataService<ImageFile> {
     return this.getOnce(imageId);
   }
   async saveNewImage(imageSaveInfo: ImageSaveInfo, id?: string) {
-    const key = id || this.newKey();
+    const key = id || await this.newKey();
     // const folder = this.activeStoreService.activeStoreKey;
     const folder = await this.activeStoreService.getlatest();
     const ImgSaveInfo = await this.imageService.extendImageSaveInfo(
@@ -59,51 +59,77 @@ export class ImagesDataService extends StoreDataService<ImageFile> {
       folder
     );
     ImgSaveInfo.uploadRes = uploadRes;
-
-//    if (uploadRes.imageTask) {
-      if (uploadRes.thumbTask) {
-        forkJoin(
-        uploadRes.imageTask.snapshotChanges(),
-        uploadRes.thumbTask.snapshotChanges()
-      )
-        .pipe(
-          switchMap(any => {
-            return combineLatest(
-              ImgSaveInfo.imageRef.getDownloadURL(),
-              ImgSaveInfo.thumbRef.getDownloadURL()
-            );
-          }),
-          catchError(err => {
-            console.log("Error upload image", err);
-            throw err;
-          })
-        )
-        // todo:
-        .subscribe(([url, thumbUrl]) => {
-          return this.afterImageUploaded(extImage, { url, thumbUrl }).then(() => {
-            this.pendingUploads.delete(extImage.id);
-          });
+    imageSaveInfo.downloadUrl = uploadRes.imageTask.snapshotChanges().pipe(
+      last(),
+      switchMap(() => imageSaveInfo.imageRef.getDownloadURL()),
+      switchMap((url) => {
+        return this.afterImageUploaded(extImage, { url, thumbUrl: null }).then(() => {
+          URL.revokeObjectURL(imageSaveInfo.editedBlobUrl);
+          this.pendingUploads.delete(extImage.id);
+          return url;
         });
-    } else {
-      uploadRes.imageTask.snapshotChanges().toPromise().then(() => {
-        return ImgSaveInfo.imageRef.getDownloadURL()
+      })
+    );
 
-          // todo:
-          .subscribe((url) => {
-            return this.afterImageUploaded(extImage, { url, thumbUrl: null }).then(() => {
-              URL.revokeObjectURL(uploadRes.imgSaveInfo.editedBlobUrl);
-              this.pendingUploads.delete(extImage.id);
+    //    if (uploadRes.imageTask) {
+    // if (uploadRes.thumbTask) {
+    //   forkJoin(
+    //     uploadRes.imageTask.snapshotChanges(),
+    //     uploadRes.thumbTask.snapshotChanges()
+    //   )
+    //     .pipe(
+    //       switchMap(any => {
+    //         return combineLatest(
+    //           ImgSaveInfo.imageRef.getDownloadURL(),
+    //           ImgSaveInfo.thumbRef.getDownloadURL()
+    //         );
+    //       }),
+    //       catchError(err => {
+    //         console.log("Error upload image", err);
+    //         throw err;
+    //       })
+    //     )
+    //     // todo:
+    //     .subscribe(([url, thumbUrl]) => {
+    //       return this.afterImageUploaded(extImage, { url, thumbUrl }).then(() => {
+    //         this.pendingUploads.delete(extImage.id);
+    //       });
+    //     });
+    // } else {
+    // return uploadRes.imageTask.snapshotChanges().toPromise().then(() => {
+    // uploadRes.imageTask.snapshotChanges()
+    //   .pipe(
+    //     finalize(() => {
+    // //      ImgSaveInfo.downloadUrl = imageSaveInfo.imageRef.getDownloadURL();
+    // ImgSaveInfo.downloadUrl.subscribe();
+    //   })
+    // )
+    // .subscribe();
+    // ImgSaveInfo.imageRef.getDownloadURL()
 
-            });
-          },
-            err => {
-              console.log("Error upload image", err);
-              throw err;
-            }
-          );
-      });
+    //   // todo:
+    //   .pipe(map(data => data as string), tap(((url) => {
+    //     // ImgSaveInfo.downloadUrl = url;
+    //     return this.afterImageUploaded(extImage, { url, thumbUrl: null }).then(() => {
+    //       URL.revokeObjectURL(uploadRes.imgSaveInfo.editedBlobUrl);
+    //       this.pendingUploads.delete(extImage.id);
 
-    }
+    //     });
+    //   }),
+    //     catchError(
+    //       err => {
+    //         console.log("Error upload image", err);
+    //         throw err;
+    //       }
+    //     )
+    //   )).subscribe(
+    //     (res) => {
+    //       console.log(res);
+    //     }
+    //   );
+    // });
+    // }
+    return ImgSaveInfo;
   }
 
   remove(removedItem: Extended<ImageFile>) {
@@ -141,6 +167,6 @@ export class ImagesDataService extends StoreDataService<ImageFile> {
 
   private afterImageUploaded(newImage: Extended<ImageFile>, { url, thumbUrl }) {
     newImage.data = { ...newImage.data, url, thumbUrl, isUploaded: true };
-    return super.saveOld(newImage);
+    return super.saveOld(newImage).then(() => this.getOnce(newImage.id));
   }
 }
